dom/imptests/html/html/dom/documents/dom-tree-accessors/test_nameditem-01.html

<!DOCTYPE html>
<title>Named items: img id &amp; name</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-nameditem">
<link rel="stylesheet" href="/resources/testharness.css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="test">
<img id="a" name="b">
</div>
<script>
test(function() {
  assert_equals(document.a, document.getElementsByTagName("img")[0]);
  assert_equals(document['a'], document.getElementsByTagName("img")[0]);
  assert_equals(document.b, document.getElementsByTagName("img")[0]);
  assert_equals(document['b'], document.getElementsByTagName("img")[0]);
});
</script>
